requirements.txt
Flask
opencv-python
Pillow
torch

